feat: implement user authentication and authorization with Flask-Login and Bcrypt

+++ app.py
@@ -1,5 +1,7 @@
 from flask import Flask, render_template, url_for, flash, redirect, request, abort
-from models import db, Post
+from models import db, Post, User
+from flask_bcrypt import Bcrypt
+from flask_login import LoginManager, login_user, current_user, logout_user, login_required
 import os
 
 app = Flask(__name__)
@@ -8,6 +10,14 @@ app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 
 db.init_app(app)
+bcrypt = Bcrypt(app)
+login_manager = LoginManager(app)
+login_manager.login_view = 'login'
+login_manager.login_message_category = 'info'
+
+@login_manager.user_loader
+def load_user(user_id):
+    return User.query.get(int(user_id))
 
 # Create the database within the application context
 with app.app_context():
@@ -19,17 +29,59 @@ def home():
     posts = Post.query.order_by(Post.date_posted.desc()).all()
     return render_template('index.html', posts=posts)
 
+@app.route("/register", methods=['GET', 'POST'])
+def register():
+    if current_user.is_authenticated:
+        return redirect(url_for('home'))
+    if request.method == 'POST':
+        username = request.form.get('username')
+        email = request.form.get('email')
+        password = request.form.get('password')
+        confirm_password = request.form.get('confirm_password')
+        
+        if password != confirm_password:
+            flash('Passwords do not match!', 'danger')
+        else:
+            hashed_password = bcrypt.generate_password_hash(password).decode('utf-8')
+            user = User(username=username, email=email, password=hashed_password)
+            db.session.add(user)
+            db.session.commit()
+            flash('Your account has been created! You are now able to log in', 'success')
+            return redirect(url_for('login'))
+    return render_template('register.html', title='Register')
+
+@app.route("/login", methods=['GET', 'POST'])
+def login():
+    if current_user.is_authenticated:
+        return redirect(url_for('home'))
+    if request.method == 'POST':
+        email = request.form.get('email')
+        password = request.form.get('password')
+        user = User.query.filter_by(email=email).first()
+        if user and bcrypt.check_password_hash(user.password, password):
+            login_user(user, remember=request.form.get('remember'))
+            next_page = request.args.get('next')
+            return redirect(next_page) if next_page else redirect(url_for('home'))
+        else:
+            flash('Login Unsuccessful. Please check email and password', 'danger')
+    return render_template('login.html', title='Login')
+
+@app.route("/logout")
+def logout():
+    logout_user()
+    return redirect(url_for('home'))
+
 @app.route("/post/new", methods=['GET', 'POST'])
+@login_required
 def new_post():
     if request.method == 'POST':
         title = request.form.get('title')
         content = request.form.get('content')
-        author = request.form.get('author') or 'Anonymous'
         
         if not title or not content:
             flash('Title and Content are required!', 'danger')
         else:
-            post = Post(title=title, content=content, author=author)
+            post = Post(title=title, content=content, author=current_user)
             db.session.add(post)
             db.session.commit()
             flash('Your post has been created!', 'success')
@@ -43,12 +95,14 @@ def post(post_id):
     return render_template('post_detail.html', title=post.title, post=post)
 
 @app.route("/post/<int:post_id>/update", methods=['GET', 'POST'])
+@login_required
 def update_post(post_id):
     post = Post.query.get_or_404(post_id)
+    if post.author != current_user:
+        abort(403)
     if request.method == 'POST':
         post.title = request.form.get('title')
         post.content = request.form.get('content')
-        post.author = request.form.get('author') or 'Anonymous'
         
         if not post.title or not post.content:
             flash('Title and Content are required!', 'danger')
@@ -60,8 +114,11 @@ def update_post(post_id):
     return render_template('create_post.html', title='Update Post', legend='Update Post', post=post)
 
 @app.route("/post/<int:post_id>/delete", methods=['POST'])
+@login_required
 def delete_post(post_id):
     post = Post.query.get_or_404(post_id)
+    if post.author != current_user:
+        abort(403)
     db.session.delete(post)
     db.session.commit()
     flash('Your post has been deleted!', 'success')
